+++ tests/utils.c
@@ -5,7 +5,7 @@
 START_TEST(test_rotation)
 {
 	unsigned int i32 = 0x11223344;
-	unsigned long i64 = 0x1122334455667788;
+	unsigned long long i64 = 0x1122334455667788;
 
 	i32 = ROTR32(i32, 8);
 	ck_assert_int_eq(0x44112233, i32);
@@ -18,7 +18,7 @@ START_TEST(test_get)
 {
 	unsigned char a[] = {0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, 0x88};
 	unsigned int i32 = GETU32_BE(a);
-	unsigned long i64 = GETU64_BE(a);
+	unsigned long long i64 = GETU64_BE(a);
 
 	ck_assert_int_eq(0x11223344, i32);
 	ck_assert_int_eq(0x1122334455667788, i64);
